Disband BrickTestHelper
Fixes https://github.com/wayfair/brickkit-android/issues/191
